+++ src/App.js
@@ -1,10 +1,10 @@
-import TitleSearchBar from "./UI/TitleSearchBar/TitleSearchBar";
-import Thumbnail from "./UI/Thumbnail/Thumbnail";
-import MovieListSection from "./UI/MovieListSection/MovieListSection";
-import Button from "./UI/Button/Button";
+import TitleSearchBar from "./UI/Header/TitleSearchBar/TitleSearchBar";
+import Thumbnail from "./UI/Header/Thumbnail/Thumbnail";
+import Button from "./UI/MoviesContainer/Button/Button";
 import Footer from "./UI/Footer/Footer";
 import { getPopularityMovie } from "./Domain/getPopularityMovie";
 import { searchMovie } from "./Domain/searchMovie";
+import MovieListSection from "./UI/MoviesContainer/MovieListSection/MovieListSection";
 
 class App {
   #movies;
