+++ src/UI/Header/Thumbnail/Thumbnail.css
