+++ src/UI/Header/Thumbnail/Thumbnail.js
