+++ src/UI/MoviesContainer/EmptyView/EmptyView.js
