+++ src/UI/MoviesContainer/MovieItem/MovieItem.css
