+++ src/UI/MoviesContainer/MovieItem/MovieItem.js
