+++ src/UI/MoviesContainer/MovieListSection/MovieListSection.css
